+++ packages/signers/signer/src/from-raw-signer.ts
@@ -1,4 +1,5 @@
 import {
+  Binary,
   Blake2256,
   V15,
   compact,
@@ -21,12 +22,16 @@ const signingTypeId: Record<"Ecdsa" | "Ed25519" | "Sr25519", number> = {
   Ecdsa: 2,
 }
 
+const [preBytes, postBytes] = ["<Bytes>", "</Bytes>"].map((str) =>
+  Binary.fromText(str).asBytes(),
+)
+
 export function getPolkadotSigner(
   publicKey: Uint8Array,
   signingType: "Ecdsa" | "Ed25519" | "Sr25519",
   sign: (input: Uint8Array) => Promise<Uint8Array> | Uint8Array,
 ): PolkadotSigner {
-  const polkadotSign = async (
+  const signTx = async (
     callData: Uint8Array,
     signedExtensions: Record<
       string,
@@ -76,5 +81,25 @@ export function getPolkadotSigner(
     return mergeUint8(compact.enc(preResult.length), preResult)
   }
 
-  return { publicKey, sign: polkadotSign }
+  const signBytes = async (data: Uint8Array) => {
+    let isPadded = true
+    let i: number
+
+    for (i = 0; isPadded && i < preBytes.length; i++)
+      isPadded = preBytes[i] === data[i]
+    isPadded = isPadded && i === preBytes.length
+
+    const postDataStart = data.length - postBytes.length
+    for (i = 0; isPadded && i < postBytes.length; i++)
+      isPadded = postBytes[i] === data[postDataStart + i]
+    isPadded = isPadded && i === postBytes.length
+
+    return sign(isPadded ? data : mergeUint8(preBytes, data, postBytes))
+  }
+
+  return {
+    publicKey,
+    signTx,
+    signBytes,
+  }
 }