+++ packages/signers/signer/tests/signBytes.spec.ts
@@ -1,0 +1,35 @@
+import { expect, describe, it } from "vitest"
+import { getPolkadotSigner } from "@/from-raw-signer"
+import { fromHex, toHex } from "@polkadot-api/utils"
+import { ed25519 } from "@noble/curves/ed25519"
+import { Binary } from "@polkadot-api/substrate-bindings"
+
+const aliceEd25519PrivKey = fromHex(
+  "0xabf8e5bdbe30c65656c0a3cbd181ff8a56294a69dfedd27982aace4a76909115",
+)
+
+const signer = getPolkadotSigner(
+  ed25519.getPublicKey(aliceEd25519PrivKey),
+  "Ed25519",
+  (input) => ed25519.sign(input, aliceEd25519PrivKey),
+)
+
+describe("signBytes", () => {
+  it("provides the same signature for padded/unpadded", async () => {
+    const unpadded = Binary.fromText("hello world").asBytes()
+    const padded = Binary.fromText("<Bytes>hello world</Bytes>").asBytes()
+
+    expect(toHex(await signer.signBytes(unpadded))).toBe(
+      toHex(await signer.signBytes(padded)),
+    )
+  })
+
+  it("provides the same signature for empty padded/unpadded", async () => {
+    const unpadded = Binary.fromText("").asBytes()
+    const padded = Binary.fromText("<Bytes></Bytes>").asBytes()
+
+    expect(toHex(await signer.signBytes(unpadded))).toBe(
+      toHex(await signer.signBytes(padded)),
+    )
+  })
+})