+++ packages/client/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - re-export `getSs58AddressInfo` from `@polkadot-api/substrate-bindings`
 
+### Changed
+
+- `polkadot-signer`: Add API to sign raw-data and rename `sign` to `signTx`
+
 ## 0.9.1 - 2024-07-03
 
 ### Fixed